requirements.txt
Django
requests
dateconverter~=0.1
elasticsearch==7.15.2
PyMuPdf
django-ckeditor
redis
beautifulsoup4
djangorestframework